# Fix benchmark_stim.py not asking for bit packed data