--- a/example/benchmark_stim.py
+++ b/example/benchmark_stim.py
@@ -3,7 +3,7 @@ from time import time
 
 circuit = stim.Circuit.from_file(f"./stim_benchmark/random10.stim")
 sampler = circuit.compile_sampler()
-sampler.sample(shots=10000)
+sampler.sample(shots=10000, bit_packed=True)
 
 with open("./benchmark_stim.dat", "w") as io:
     print("nq init_time sample_time", file=io)
@@ -12,7 +12,7 @@ with open("./benchmark_stim.dat", "w") as io:
         circuit = stim.Circuit.from_file(f"./stim_benchmark/random{j}.stim")
         sampler = circuit.compile_sampler()
         t1 = time()
-        sampler.sample(shots=10000)
+        sampler.sample(shots=10000, bit_packed=True)
         t2 = time()
         print(f"{j} {t1-t0} {t2-t1}", file=io)
         print(f"{j} {t1-t0} {t2-t1}")