@njuwfang 2023-11-21
Sorry that I ignored detector_error_model in Stim. 

I find that https://github.com/quantumlib/Stim/blob/main/src/stim/simulators/error_analyzer.h obtains a detector_error_model in reverse order.
It seems that Stim uses the same method as the recent [ABC sim.](https://arxiv.org/abs/2309.15345)?